test(generate tool): setup, cases
Created an initial setup for a `bats` test harness see:
[bats](https://github.com/bats-core).

First test covering the script is not run within a git repository.

Install `bats` from git in the docker image instead of relying on an apk
package, as recommended in the documentation.

Added tests for generate when it's used outside of a git repository and
when the git repository does not contain any commit

generate fails if there is not any conventional commit found in the
repository