test(generate): outside-in baby!
Preparing a new test suite for some algorithm that are implementation
details of the more general changelog generation algorithm. This suite
is designed to test drive complex inner behavior for changelog
generation.

A first test to handle Unreleased section.
It was a P.I.T.A., really to make this one.
Need refacto regarding the way to count multi line output line count