refactor(tests): all in docker
All generate calls are now in the docker in docker envrionment. Ensure
real conformance with the main application in real world that will take
place within a docker container. Even if it is not the case
(`--no-docker` option set) executing generates.sh outside of a docker
container is a subset of executing generate.sh within a container.