feat(generate): from ancestor limit
generate.sh now accepts the `--from-ancestor` option. Specify a revision
value to this option that corresponds to the ancestor of the commit you
want to start at for the changelog generation. As stated in the help,
this is useful if you work with a fork of a repository and want to
redefine the versioning scheme.